Refactored code into individual services
